maintenance: Implement missing remove_scores script in Python.
There used to be a maintenance script, which had not been added to the
repository.  Instead of calling the script we do the modification in
Python.  Unfortunately, the script has been lost, so we don't know
exactly what it did.  Comparing the generated hub files, however, this
appears to be equivalent to whatever the script did.